Updating ai.py
